[#11] new contract for running function
This commits changes the way we run our function in the cloud, before we
passed the args and a callback function, now a req object is
deliveried with the basic content that is needed and a res object. One
can set the response, add headers and change the status code of the
response.