feat: add support for removing protected fields from response
The SchemaResponse class now has a removeProtectedFields method that removes fields specified in the responseProtectedFields configuration option from the response data. The responseProtectedFields configuration option is a list of fields that should be removed from the response data. This allows sensitive information to be removed from the response before it is sent to the client.